--- cartridges/custom_sanika1/cartridge/templates/default/checkout/shipping/shippingCard.isml
@@ -1,36 +1,0 @@
-<isif condition="${shippingModel.shippingAddress}">
-    <h3>${Resource.msg('heading.shipping.to', 'checkout', null)}</h3>
-    <isinclude template="checkout/shipping/shippingCardShippingAddress" />
-</isif>
-
-<isif condition="${shippingModel.selectedShippingMethod}">
-    <h3>${pdict.order.resources.shippingMethod}
-        <span>
-            <isif condition="${shippingModel.productLineItems.items && shippingModel.productLineItems.items.length > 1}">
-            ${shippingModel.productLineItems.items.length.toFixed(0)} ${pdict.order.resources.items}
-            </isif>
-        </span>
-    </h3>
-    <div class="row leading-lines">
-        <p class="col-8 start-lines">
-            <span class="display-name">
-                ${shippingModel.selectedShippingMethod && shippingModel.selectedShippingMethod.displayName ?
-                    shippingModel.selectedShippingMethod.displayName : 'No Shipping Method Selected'}
-            </span>
-            <isif condition="${shippingModel.selectedShippingMethod.estimatedArrivalTime}">
-                <span class="text-muted arrival-time">
-                    ${shippingModel.selectedShippingMethod.estimatedArrivalTime ?
-                        shippingModel.selectedShippingMethod.estimatedArrivalTime : 'No Estimated Arrival'}
-                </span>
-            </isif>
-        </p>
-        <p class="col-4 end-lines text-right">
-            <span class="price">
-                ${shippingModel.selectedShippingMethod && shippingModel.selectedShippingMethod.shippingCost ?
-                    shippingModel.selectedShippingMethod.shippingCost : 'No Shipping Cost'}
-            </span>
-        </p>
-    </div>
-</isif>
-
-<isinclude template="checkout/shipping/giftMessageSummary" />
